requirements.txt
numpy>=1.26.0
Pillow>=10.0.0
opencv-contrib-python-headless>=4.10.0
ultralytics>=8.3.0
pillow-avif-plugin>=1.4.6
